completely restructured the navigation of the docs
